pyproject.toml
[project]
name = "prarielearn-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "notion-client>=2.7.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "pytz>=2025.2",
    "selenium>=4.39.0",
]
